+++ gouda/barcode.py
@@ -1,7 +1,7 @@
 import collections
 
 
-class Barcode(collections.namedtuple('Barcode', ['type', 'data'])):
+class Barcode(collections.namedtuple('Barcode', ['type', 'data', 'rect'])):
     """A simple representation of a barcode
     """
     pass