+++ gouda/engines/datasymbol.py
@@ -69,7 +69,10 @@ class DataSymbolEngine(object):
         res = [None] * self.d.Barcodes.length
         for i in range(0, self.d.Barcodes.length):
             b = self.d.Barcodes.item(i)
-            res[i] = Barcode(self.types.get(b.BarcodeType, 'Unknown'), b.Text)
+            # TODO Coordinates
+            res[i] = Barcode(
+                self.types.get(b.BarcodeType, 'Unknown'), b.Text, None
+            )
         return res
 
     def __call__(self, img):