+++ gouda/strategies/resize.py
@@ -1,6 +1,8 @@
 import cv2
 import numpy as np
 
+from gouda.barcode import Barcode
+from gouda.rect import Rect
 from gouda.util import debug_print
 
 
@@ -27,6 +29,18 @@ def resize(img, engine):
                 # Resize from the original image
                 i = cv2.resize(img, (0, 0), fx=f, fy=f)
             barcodes = engine(i)
+
+            # Adjust coordinates to account for resizing
+            for index, barcode in enumerate(barcodes):
+                if barcode.rect:
+                    barcodes[index] = Barcode(
+                        barcode.type,
+                        barcode.data,
+                        Rect(
+                            *map(lambda v: int(v / f), barcode.rect)
+                        )
+                    )
+
             if barcodes:
                 return msg, barcodes